Move views from users to core
