+++ src/core/jinja2/core/components/artworks-list.jinja2
@@ -3,7 +3,7 @@
 {% block content %}
 
         {% with repository_list = repository_list, element_type="artwork" %}
-            {% include "users/components/item-list.jinja2" %}
+            {% include "core/components/item-list.jinja2" %}
         {% endwith%}
 
         <script>
