+++ src/core/jinja2/core/components/item-list.jinja2
