+++ src/core/jinja2/core/components/repository-list.jinja2
@@ -3,7 +3,7 @@
 {% block content %}
 
         {% with repository_list = repository_list, element_type=element_type %}
-            {% include "users/components/item-list.jinja2" %}
+            {% include "core/components/item-list.jinja2" %}
         {% endwith%}
 
         <script>
