+++ src/core/jinja2/core/exhibit_create.jinja2
@@ -36,7 +36,7 @@
                     {% endif %}
 
                     {% with repository_list = artworks, element_type="artwork", selected=selected_artworks %}
-                        {% include "users/components/artworks-list.jinja2" %}
+                        {% include "core/components/artworks-list.jinja2" %}
                     {% endwith%}
 
                     <p id="artworks-selected">
