+++ src/core/jinja2/core/upload-object.jinja2
@@ -54,7 +54,7 @@
                         </span>
 
                         {% with xposition=0, yposition=0 %}
-                            {% include "users/components/object-position.jinja2" %}
+                            {% include "core/components/object-position.jinja2" %}
                         {% endwith %}
                         
                         {{ form.hidden_fields()[0] }}
