+++ src/core/urls.py
@@ -7,38 +7,48 @@ from core.views.static_views import (
     favicon,
     health_check,
     home,
+    manifest,
     marker_generator,
     robots_txt,
+    service_worker,
 )
 from core.views.views import (
     artwork_preview,
     collection,
+    create_artwork,
     create_exhibit,
+    delete,
+    edit_artwork,
     edit_exhibit,
+    edit_marker,
     edit_object,
     exhibit,
     exhibit_detail,
     exhibit_select,
-    manifest,
+    marker_upload,
     object_upload,
     related_content,
     see_all,
-    service_worker,
 )
 
 urlpatterns = [
     path("", home, name="home"),
-    path("documentation/", documentation, name="documentation"),
-    path("community/", community, name="community"),
+    path("artwork/", artwork_preview, name="artwork-preview"),
+    path("artworks/create/", create_artwork, name="create-artwork"),
+    path("artworks/edit/", edit_artwork, name="edit-artwork"),
     path("collection/", collection, name="collection"),
+    path("community/", community, name="community"),
+    path("content/delete/", delete, name="delete-content"),
+    path("documentation/", documentation, name="documentation"),
     path("exhibit_select/", exhibit_select, name="exhibit_select"),
     path("exhibit/", exhibit_detail, name="exhibit-detail"),
     path("exhibits/create/", create_exhibit, name="create-exhibit"),
     path("exhibits/edit/", edit_exhibit, name="edit-exhibit"),
-    path("artwork/", artwork_preview, name="artwork-preview"),
-    path("objects/upload/", object_upload, name="object-upload"),
-    path("objects/edit/", edit_object, name="edit-object"),
     path("generator/", marker_generator, name="marker-generator"),
+    path("markers/edit/", edit_marker, name="edit-marker"),
+    path("markers/upload/", marker_upload, name="marker-upload"),
+    path("objects/edit/", edit_object, name="edit-object"),
+    path("objects/upload/", object_upload, name="object-upload"),
     path("related-content", related_content, name="related-content"),
     path("sw.js", service_worker, name="sw"),
     path("manifest.json", manifest, name="manifest"),
