+++ src/users/views.py
@@ -16,9 +16,6 @@ from django.urls import reverse_lazy
 from django.utils.translation import gettext_lazy as _
 from django.views.decorators.http import require_http_methods
 
-from core.forms import ArtworkForm, UploadMarkerForm, UploadObjectForm
-from core.models import Artwork, Exhibit, Marker, Object
-
 from .forms import (
     PasswordChangeForm,
     ProfileForm,
@@ -115,212 +112,6 @@ def profile(request):
     return render(request, "users/profile.jinja2", ctx)
 
 
-def get_element(request, form_class, form_type, source, author, existent_element):
-    element = None
-
-    if source and author:
-        instance = form_type(source=source, author=author)
-        element = form_class(instance=instance).save(commit=False)
-        element.save()
-    elif existent_element:
-        qs = form_type.objects.filter(id=existent_element)
-        if qs:
-            element = qs[0]
-            element.owner = request.user.profile
-
-    return element
-
-
-def get_marker(request, form):
-    marker_src = form.cleaned_data["marker"]
-    marker_author = form.cleaned_data["marker_author"]
-    existent_marker = form.cleaned_data["existent_marker"]
-
-    return get_element(
-        request,
-        UploadMarkerForm,
-        Marker,
-        source=marker_src,
-        author=marker_author,
-        existent_element=existent_marker,
-    )
-
-
-def get_augmented(request, form):
-    object_src = form.cleaned_data["augmented"]
-    object_author = form.cleaned_data["augmented_author"]
-    existent_object = form.cleaned_data["existent_object"]
-
-    return get_element(
-        request,
-        UploadObjectForm,
-        Object,
-        source=object_src,
-        author=object_author,
-        existent_element=existent_object,
-    )
-
-
-@login_required
-def create_artwork(request):
-    if request.method == "POST":
-        form = ArtworkForm(request.POST, request.FILES)
-
-        if form.is_valid():
-            marker = get_marker(request, form)
-            augmented = get_augmented(request, form)
-
-            if marker and augmented:
-                artwork_title = form.cleaned_data["title"]
-                artwork_desc = form.cleaned_data["description"]
-                Artwork(
-                    author=request.user.profile,
-                    marker=marker,
-                    augmented=augmented,
-                    title=artwork_title,
-                    description=artwork_desc,
-                ).save()
-            return redirect("profile")
-    else:
-        form = ArtworkForm()
-
-    marker_list = Marker.objects.all()
-    object_list = Object.objects.all()
-
-    return render(
-        request,
-        "core/upload-artwork.jinja2",
-        {
-            "form": form,
-            "marker_list": marker_list,
-            "object_list": object_list,
-        },
-    )
-
-
-def upload_elements(request, form_class, form_type, route):
-    if request.method == "POST":
-        form = form_class(request.POST, request.FILES)
-        if form.is_valid():
-            upload = form.save(commit=False)
-            upload.owner = request.user.profile
-            upload.save()
-            return redirect("profile")
-    else:
-        form = form_class()
-
-    if form_type == "marker":
-        return render(
-            request,
-            "core/upload-marker.jinja2",
-            {"form_type": form_type, "form": form, "route": route, "edit": False},
-        )
-
-    return render(
-        request,
-        "core/upload-object.jinja2",
-        {"form": form, "route": route, "edit": False},
-    )
-
-
-@login_required
-def marker_upload(request):
-    return upload_elements(request, UploadMarkerForm, "marker", "marker-upload")
-
-
-def edit_elements(request, form_class, route, model, model_data):
-    if not model or model.owner != Profile.objects.get(user=request.user):
-        raise Http404
-
-    if request.method == "POST":
-        form = form_class(request.POST, request.FILES, instance=model)
-
-        form.full_clean()
-        if form.is_valid():
-            form.save()
-            return redirect("profile")
-
-        log.warning(form.errors)
-
-    return render(
-        request,
-        route,
-        {
-            "form": form_class(initial=model_data),
-            "model": model,
-        },
-    )
-
-
-@login_required
-def edit_marker(request):
-    index = request.GET.get("id", "-1")
-    model = Marker.objects.get(id=index)
-
-    model_data = {
-        "source": model.source,
-        "created": model.created,
-        "author": model.author,
-        "patt": model.patt,
-        "title": model.title,
-    }
-
-    return edit_elements(
-        request,
-        UploadMarkerForm,
-        route="users/edit-marker.jinja2",
-        model=model,
-        model_data=model_data,
-    )
-
-
-@login_required
-def edit_artwork(request):
-    index = request.GET.get("id", "-1")
-    model = Artwork.objects.filter(id=index).order_by("-id")
-    if not model or model.first().author != Profile.objects.get(user=request.user):
-        raise Http404
-
-    if request.method == "POST":
-        form = ArtworkForm(request.POST, request.FILES)
-
-        form.full_clean()
-        if form.is_valid():
-            model_data = {
-                "marker": get_marker(request, form),
-                "augmented": get_augmented(request, form),
-                "title": form.cleaned_data["title"],
-                "description": form.cleaned_data["description"],
-            }
-            print(model_data["augmented"])
-            model.update(**model_data)
-            return redirect("profile")
-
-    model = model.first()
-    model_data = {
-        "marker": model.marker,
-        "marker_author": model.marker.author,
-        "augmented": model.augmented,
-        "augmented_author": model.augmented.author,
-        "title": model.title,
-        "description": model.description,
-        "existent_marker": model.marker.id,
-        "existent_object": model.augmented.id,
-    }
-
-    return render(
-        request,
-        "core/upload-artwork.jinja2",
-        {
-            "form": ArtworkForm(initial=model_data),
-            "marker_list": Marker.objects.all(),
-            "object_list": Object.objects.all(),
-            "selected_marker": model.marker.id,
-            "selected_object": model.augmented.id,
-        },
-    )
-
-
 @login_required
 def edit_password(request):
     if request.method == "POST":
@@ -365,36 +156,3 @@ def edit_profile(request):
             "form_password": PasswordChangeForm(request.user),
         },
     )
-
-
-@login_required
-@require_http_methods(["GET"])
-def delete(request):
-    content_type = request.GET.get("content_type", None)
-
-    if content_type == "marker":
-        delete_content(Marker, request.user, request.GET.get("id", -1))
-    elif content_type == "object":
-        delete_content(Object, request.user, request.GET.get("id", -1))
-    elif content_type == "artwork":
-        delete_content(Artwork, request.user, request.GET.get("id", -1))
-    elif content_type == "exhibit":
-        delete_content(Exhibit, request.user, request.GET.get("id", -1))
-    return redirect("profile")
-
-
-def delete_content(model, user, instance_id):
-    qs = model.objects.filter(id=instance_id)
-
-    if qs:
-        instance = qs[0]
-
-        is_artwork = isinstance(instance, Artwork)
-        if is_artwork:
-            has_permission = instance.author == user.profile
-        else:
-            has_permission = instance.owner == user.profile
-
-        is_instance_same_type_of_model = isinstance(instance, model)
-        if is_instance_same_type_of_model and has_permission:
-            instance.delete()
